+++ packages/preview-middleware-client/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @sap-ux-private/preview-middleware-client
 
+## 1.2.10
+
 ## 1.2.9
 
 ## 1.2.8